% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarypscl.R
\name{print.hurdle}
\alias{print.hurdle}
\title{Shorter version of hurdle model output}
\usage{
\method{print}{hurdle}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class "hurdle", usually, a result of a call to hurdle().}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Shorter version of hurdle model output
}
