% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMASS.R
\name{print.lda}
\alias{print.lda}
\title{Replaces default lda version by removing call info}
\usage{
\method{print}{lda}(x, ...)
}
\arguments{
\item{x}{an object of class "lda", usually, a result of a call to lda().}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default lda version by removing call info
}
