% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMASS.R
\name{print.polr}
\alias{print.polr}
\title{Replaces default polr version by removing call info}
\usage{
\method{print}{polr}(x, ...)
}
\arguments{
\item{x}{an object of class "polr", usually, a result of a call to polr().}

\item{...}{further arguments passed to or from other methods.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default polr version by removing call info
}
