% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarycoxph.R
\name{print.summary.coxph}
\alias{print.summary.coxph}
\title{Shorter coxph model summary output}
\usage{
\method{print}{summary.coxph}(
  x,
  digits = max(getOption("digits") - 3, 3),
  signif.stars = getOption("show.signif.stars"),
  expand = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class "summary.coxph", usually, a result of a call to summary.coxph.}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{expand}{logical. If the summary is for a multi-state coxph fit, print the results in an expanded format.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Shorter coxph model summary output
}
