% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarypscl.R
\name{print.summary.hurdle}
\alias{print.summary.hurdle}
\title{Replaces default hurdle version by removing useless call and residuals info}
\usage{
\method{print}{summary.hurdle}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class "summary.hurdle", usually, a result of a call to summary().}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default hurdle version by removing useless call and residuals info
}
