% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarylmer.R
\name{print.summary.merMod}
\alias{print.summary.merMod}
\title{Shorter lmer model summary output}
\usage{
\method{print}{summary.merMod}(
  x,
  digits = max(3, getOption("digits") - 3),
  correlation = FALSE,
  symbolic.cor = FALSE,
  signif.stars = getOption("show.signif.stars"),
  ranef.comp = c("Variance", "Std.Dev."),
  ranef.corr = any(ranef.comp == "Std.Dev."),
  show.resids = TRUE,
  concise = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class "summary.merMod", usually, a result of a call to summary.merMod.}

\item{digits}{the number of significant digits to use when printing.}

\item{correlation}{logical. Indicates whether the correlation matrix of the fixed-effects parameters should be printed.}

\item{symbolic.cor}{logical. Should a symbolic encoding of the fixed-effects correlation matrix be printed?}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{ranef.comp}{character vector of length one or two. Indicates if random-effects parameters should be reported on the variance and/or standard deviation scale.}

\item{ranef.corr}{logical. Print correlations (rather than covariances) of random effects?}

\item{show.resids}{logical. Should the quantiles of the scaled residuals be printed?}

\item{concise}{logical. If TRUE, don't print the call and residuals stats.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Shorter lmer model summary output
}
