% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMASS.R
\name{print.summary.polr}
\alias{print.summary.polr}
\title{Replaces default polr version by removing call info}
\usage{
\method{print}{summary.polr}(x, digits = x$digits, concise = TRUE, ...)
}
\arguments{
\item{x}{an object of class "summary.polr", usually, a result of a call to summary().}

\item{digits}{the number of significant digits to use when printing.}

\item{concise}{logical. If TRUE, don't print AIC.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default polr version by removing call info
}
