% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarypscl.R
\name{print.zeroinfl}
\alias{print.zeroinfl}
\title{Shorter version of zeroinfl model output}
\usage{
\method{print}{zeroinfl}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class "zeroinfl", usually, a result of a call to zeroinfl().}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Shorter version of zeroinfl model output
}
