% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryglm.R
\name{print.glm}
\alias{print.glm}
\title{Replaces default stats version by removing useless call info}
\usage{
\method{print}{glm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class "lm", usually, a result of a call to lm().}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default stats version by removing useless call info
}
