% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryglm.R
\name{print.summary.glm}
\alias{print.summary.glm}
\title{Print glm summary}
\usage{
\method{print}{summary.glm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  show.residuals = FALSE,
  concise = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class "summary.glm", usually, a result of a call to summary.glm.}

\item{digits}{the number of significant digits to use when printing.}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic form}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{show.residuals}{logical. If TRUE then a summary of the deviance residuals is printed at the head of the output.}

\item{concise}{logical. If TRUE, don't print the call, AIC and number of Fisher scoring iterations.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default stats version by removing call info, AIC and Fisher scoring info.
Also removes "Coefficients:" line and other empty lines
}
