% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarylm.R
\name{print.summary.lm}
\alias{print.summary.lm}
\title{Print lm summary}
\usage{
\method{print}{summary.lm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  concise = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class "summary.lm", usually, a result of a call to summary.lm.}

\item{digits}{the number of significant digits to use when printing.}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic form}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{concise}{logical. If TRUE, don't print the call and residuals stats.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Nothing - prints output only.
}
\description{
Replaces default stats version by removing useless call info and residual stats
}
